+++ project/src/test/mailbox/reader.c
@@ -1,40 +1,39 @@
-// C Program for Message Queue (Reader Process) 
-#include <stdio.h> 
-#include <sys/ipc.h> 
-#include <sys/msg.h> 
-  
-// structure for message queue 
-struct mesg_buffer { 
-    long mesg_type; 
-    char mesg_text[10]; 
+// C Program for Message Queue (Reader Process)
+#include <stdio.h>
+#include <sys/ipc.h>
+#include <sys/msg.h>
+
+// structure for message queue
+struct mesg_buffer {
+    int mesg_type;
+    char mesg_text[10];
     key_t reponse_key;
-} message; 
-  
-int main() 
-{ 
-    key_t key; 
-    int msgid; 
-  
-    // ftok to generate unique key 
-    key = ftok("progfile", 65); 
-  
-    // msgget creates a message queue 
-    // and returns identifier 
-    msgid = msgget(key, 0666 | IPC_CREAT); 
-  
-    // msgrcv to receive message 
-    int ret = msgrcv(msgid, &message, sizeof(message), 1, 0); 
-  
-    if(ret == -1){
-    	printf("Errore %d",ret);
+} message;
+
+int main() {
+    key_t key;
+    int msgid;
+
+    // ftok to generate unique key
+    key = ftok("progfile", 65);
+
+    // msgget creates a message queue
+    // and returns identifier
+    msgid = msgget(key, 0666 | IPC_CREAT);
+
+    // msgrcv to receive message
+    int ret = msgrcv(msgid, &message, sizeof(message), 1, 0);
+
+    if (ret == -1) {
+        printf("Errore %d", ret);
     }
 
-    // display the message 
-    printf("Data Received is : %s \n",  
-                    message.mesg_text); 
-  
-    // to destroy the message queue 
-    msgctl(msgid, IPC_RMID, NULL); 
-  
-    return 0; 
+    // display the message
+    printf("Data Received is : %s \n",
+           message.mesg_text);
+
+    // to destroy the message queue
+    msgctl(msgid, IPC_RMID, NULL);
+
+    return 0;
 }
